remaning searches restructured
