+++ autofit/non_linear/optimize/pyswarms/abstract.py
@@ -8,9 +8,9 @@ from autoconf import conf
 from autofit import exc
 from autofit.mapper.prior_model.abstract import AbstractPriorModel
 from autofit.non_linear.optimize.abstract_optimize import AbstractOptimizer
-from autofit.non_linear.samples import OptimizerSamples, Sample
 from autofit.non_linear.abstract_search import PriorPasser
 from autofit.non_linear.initializer import Initializer
+from autofit.non_linear.optimize.pyswarms.samples import PySwarmsSamples
 from autofit.plot import PySwarmsPlotter
 from autofit.plot.output import Output
 
@@ -208,7 +208,7 @@ class AbstractPySwarms(AbstractOptimizer):
 
     def fitness_function_from_model_and_analysis(self, model, analysis, log_likelihood_cap=None):
 
-        return PySwarmsGlobal.Fitness(
+        return AbstractPySwarms.Fitness(
             paths=self.paths,
             model=model,
             analysis=analysis,
@@ -250,233 +250,3 @@ class AbstractPySwarms(AbstractOptimizer):
 
         if should_plot("time_series"):
             plotter.time_series()
-
-
-class PySwarmsGlobal(AbstractPySwarms):
-    __identifier_fields__ = (
-        "n_particles",
-        "cognitive",
-        "social",
-        "inertia",
-    )
-
-    def __init__(
-            self,
-            name: Optional[str] = None,
-            path_prefix: Optional[str] = None,
-            unique_tag: Optional[str] = None,
-            prior_passer: Optional[PriorPasser] = None,
-            initializer: Optional[Initializer] = None,
-            iterations_per_update: int = None,
-            number_of_cores: int = None,
-            session: Optional[Session] = None,
-            **kwargs
-    ):
-        """
-        A PySwarms Particle Swarm Optimizer global non-linear search.
-
-        For a full description of PySwarms, checkout its Github and readthedocs webpages:
-
-        https://github.com/ljvmiranda921/pyswarms
-
-        https://pyswarms.readthedocs.io/en/latest/index.html
-
-        Parameters
-        ----------
-        name
-            The name of the search, controlling the last folder results are output.
-        path_prefix
-            The path of folders prefixing the name folder where results are output.
-        unique_tag
-            The name of a unique tag for this model-fit, which will be given a unique entry in the sqlite database
-            and also acts as the folder after the path prefix and before the search name.
-        prior_passer
-            Controls how priors are passed from the results of this `NonLinearSearch` to a subsequent non-linear search.
-        initializer
-            Generates the initialize samples of non-linear parameter space (see autofit.non_linear.initializer).
-        number_of_cores : int
-            The number of cores Emcee sampling is performed using a Python multiprocessing Pool instance. If 1, a
-            pool instance is not created and the job runs in serial.
-        """
-
-        super().__init__(
-            name=name,
-            path_prefix=path_prefix,
-            unique_tag=unique_tag,
-            prior_passer=prior_passer,
-            initializer=initializer,
-            iterations_per_update=iterations_per_update,
-            number_of_cores=number_of_cores,
-            session=session,
-            **kwargs
-        )
-
-        self.logger.debug("Creating PySwarms Search")
-
-    def sampler_from(self, model, fitness_function, bounds, init_pos):
-        """Get the static Dynesty sampler which performs the non-linear search, passing it all associated input Dynesty
-        variables."""
-
-        import pyswarms
-
-        options = {
-            "c1": self.config_dict_search["cognitive"],
-            "c2": self.config_dict_search["social"],
-            "w": self.config_dict_search["inertia"]
-        }
-
-        config_dict = self.config_dict_search
-        config_dict.pop("cognitive")
-        config_dict.pop("social")
-        config_dict.pop("inertia")
-
-        return pyswarms.global_best.GlobalBestPSO(
-            dimensions=model.prior_count,
-            bounds=bounds,
-            init_pos=init_pos,
-            options=options,
-            **config_dict
-        )
-
-
-class PySwarmsLocal(AbstractPySwarms):
-    __identifier_fields__ = (
-        "n_particles",
-        "cognitive",
-        "social",
-        "inertia",
-        "number_of_k_neighbors",
-        "minkowski_p_norm"
-    )
-
-    def __init__(
-            self,
-            name: Optional[str] = None,
-            path_prefix: Optional[str] = None,
-            unique_tag: Optional[str] = None,
-            prior_passer=None,
-            iterations_per_update: int = None,
-            number_of_cores: int = None,
-            session: Optional[Session] = None,
-            **kwargs
-    ):
-        """
-        A PySwarms Particle Swarm Optimizer global non-linear search.
-
-        For a full description of PySwarms, checkout its Github and readthedocs webpages:
-
-        https://github.com/ljvmiranda921/pyswarms
-
-        https://pyswarms.readthedocs.io/en/latest/index.html
-
-        Parameters
-        ----------
-        name
-            The name of the search, controlling the last folder results are output.
-        path_prefix
-            The path of folders prefixing the name folder where results are output.
-        unique_tag
-            The name of a unique tag for this model-fit, which will be given a unique entry in the sqlite database
-            and also acts as the folder after the path prefix and before the search name.
-        prior_passer
-            Controls how priors are passed from the results of this `NonLinearSearch` to a subsequent non-linear search.
-        initializer
-            Generates the initialize samples of non-linear parameter space (see autofit.non_linear.initializer).
-        number_of_cores : int
-            The number of cores Emcee sampling is performed using a Python multiprocessing Pool instance. If 1, a
-            pool instance is not created and the job runs in serial.
-        """
-
-        super().__init__(
-            name=name,
-            path_prefix=path_prefix,
-            unique_tag=unique_tag,
-            prior_passer=prior_passer,
-            iterations_per_update=iterations_per_update,
-            number_of_cores=number_of_cores,
-            session=session,
-            **kwargs
-        )
-
-        self.logger.debug("Creating PySwarms Search")
-
-    def sampler_from(self, model, fitness_function, bounds, init_pos):
-        """
-        Get the static Dynesty sampler which performs the non-linear search, passing it all associated input Dynesty
-        variables.
-        """
-
-        import pyswarms
-
-        options = {
-            "c1": self.config_dict_search["cognitive"],
-            "c2": self.config_dict_search["social"],
-            "w": self.config_dict_search["inertia"],
-            "k": self.config_dict_search["number_of_k_neighbors"],
-            "p": self.config_dict_search["minkowski_p_norm"],
-        }
-
-        config_dict = self.config_dict_search
-        config_dict.pop("cognitive")
-        config_dict.pop("social")
-        config_dict.pop("inertia")
-        config_dict.pop("number_of_k_neighbors")
-        config_dict.pop("minkowski_p_norm")
-
-        return pyswarms.local_best.LocalBestPSO(
-            dimensions=model.prior_count,
-            bounds=bounds,
-            init_pos=init_pos,
-            options=options,
-            **config_dict
-        )
-
-
-class PySwarmsSamples(OptimizerSamples):
-
-    def __init__(
-            self,
-            model: AbstractPriorModel,
-            points: np.ndarray,
-            log_posterior_list: np.ndarray,
-            total_iterations: int,
-            time: Optional[float] = None,
-    ):
-        """
-        Create an *OptimizerSamples* object from this non-linear search's output files on the hard-disk and model.
-
-        For PySwarms, all quantities are extracted via pickled states of the particle and cost histories.
-
-        Parameters
-        ----------
-        model
-            The model which generates instances for different points in parameter space. This maps the points from unit
-            cube values to physical values via the priors.
-        """
-
-        self.points = points
-        self._log_posterior_list = log_posterior_list
-        self.total_iterations = total_iterations
-
-        parameter_lists = [
-            param.tolist() for parameters in self.points for param in parameters
-        ]
-        log_prior_list = [
-            sum(model.log_prior_list_from_vector(vector=vector)) for vector in parameter_lists
-        ]
-        log_likelihood_list = [lp - prior for lp, prior in zip(self._log_posterior_list, log_prior_list)]
-        weight_list = len(log_likelihood_list) * [1.0]
-
-        sample_list = Sample.from_lists(
-            model=model,
-            parameter_lists=[parameters.tolist()[0] for parameters in self.points],
-            log_likelihood_list=log_likelihood_list,
-            log_prior_list=log_prior_list,
-            weight_list=weight_list
-        )
-
-        super().__init__(
-            model=model,
-            sample_list=sample_list,
-            time=time,
-        )
