+++ autofit/non_linear/optimize/pyswarms/plotter.py
