+++ autofit/plot/__init__.py
@@ -1,6 +1,6 @@
 from autofit.plot.samples_plotters import SamplesPlotter
-from autofit.plot.dynesty_plotter import DynestyPlotter
-from autofit.plot.ultranest_plotter import UltraNestPlotter
+from autofit.non_linear.nest.dynesty.plotter import DynestyPlotter
+from autofit.non_linear.nest.ultranest.plotter import UltraNestPlotter
 from autofit.non_linear.mcmc.emcee.plotter import EmceePlotter
 from autofit.non_linear.mcmc.zeus.plotter import ZeusPlotter
-from autofit.plot.pyswarms_plotter import PySwarmsPlotter
+from autofit.non_linear.optimize.pyswarms.plotter import PySwarmsPlotter
